refactor: clarify event editor data flow
