+++ packages/ilmomasiina-frontend/src/store/types.ts
@@ -2,7 +2,7 @@ import { Action, AnyAction, Store } from "redux";
 import { ThunkAction, ThunkDispatch } from "redux-thunk";
 
 import { AdminEventsActions } from "../modules/adminEvents/types";
-import { EditorActions } from "../modules/editor/types";
+import { EditorActions } from "../modules/editor/actions";
 import { makeRootReducer } from "./reducers";
 
 export type AppReducer = ReturnType<typeof makeRootReducer>;
